Move related-proposals to 'proposals' folder
