+++ src/ggrc/assets/javascripts/components/related-objects/proposals/related-proposals-item.js
@@ -3,12 +3,12 @@
  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 
-import '../proposal/apply-decline-proposal';
-import '../diff/instance-fields-diff';
-import '../diff/instance-acl-diff';
-import '../diff/instance-gca-diff';
-import '../diff/instance-mapping-fields-diff';
-import '../diff/instance-list-fields-diff';
+import '../../proposal/apply-decline-proposal';
+import '../../diff/instance-fields-diff';
+import '../../diff/instance-acl-diff';
+import '../../diff/instance-gca-diff';
+import '../../diff/instance-mapping-fields-diff';
+import '../../diff/instance-list-fields-diff';
 import template from './templates/related-proposals-item.mustache';
 const tag = 'related-proposals-item';
