+++ src/ggrc/assets/javascripts/components/related-objects/proposals/related-proposals.js
