+++ src/ggrc/assets/javascripts/components/related-objects/proposals/templates/related-proposals-item.mustache
