+++ src/ggrc/assets/javascripts/controllers/info_pin_controller.js
@@ -17,8 +17,8 @@ import '../components/object-list-item/document-object-list-item';
 import '../components/object-list-item/editable-document-object-list-item';
 import '../components/show-related-assessments-button/show-related-assessments-button';
 import '../components/proposal/create-proposal-button';
-import '../components/related-objects/related-proposals';
-import '../components/related-objects/related-proposals-item';
+import '../components/related-objects/proposals/related-proposals';
+import '../components/related-objects/proposals/related-proposals-item';
 import '../components/unarchive_link';
 import '../components/sort/sort-by';
 import * as TreeViewUtils from '../plugins/utils/tree-view-utils';
